putting items data to app js as root and using props to reuseing items data in footer and content as sibling
