+++ src/Content.js
@@ -1,40 +1,8 @@
 import React, { Component } from "react";
-import { useState } from "react";
 import { FaTrashAlt } from "react-icons/fa";
 
-const Content = () => {
-  const [items, setItems] = useState([
-    {
-      id: 1,
-      checked: true,
-      item: "One half pound bag of Cocoa Covered Almonds Unsalted",
-    },
-    {
-      id: 2,
-      checked: false,
-      item: "Item 2",
-    },
-    {
-      id: 3,
-      checked: false,
-      item: "Item 3",
-    },
-  ]);
-
-  const handleCheck = (id) => {
-    const listItems = items.map((item) =>
-      item.id === id ? { ...item, checked: !item.checked } : item
-    );
-    setItems(listItems);
-    localStorage.setItem('shoppingList' , JSON.stringify(listItems));
-  };
-
-  const handleDelete = (id) => {
-    const listItems = items.filter((item) => item.id !== id) ;
-    setItems(listItems);
-    localStorage.setItem('shoppingList' , JSON.stringify(listItems))
-  }
-
+const Content = ({items, handleCheck, handleDelete}) => {
+  
   return (
     <main>
       {items.length ? (
