+++ src/Footer.js
@@ -1,11 +1,10 @@
 import React from 'react'
 
-function Footer() {
-    const today = new Date();
+function Footer({length}) {
   return (
     <footer>
         <p>
-            Copyright &copy; {today.getFullYear()}
+            {length} List {length === 1 ? 'item' : 'items'}
         </p>
     </footer>
   )
